safari-600.7-branch/LayoutTests/html5lib/generated/run-tests11-write-expected.txt

CONSOLE MESSAGE: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: Warning: Problem parsing viewBox=""
../resources/tests11.dat: PASS
